Various changes and additions.
- Change every long (used during early testing) to int.

- Add Cloak, Long Staff, Robes, and Sandals.

- Change some starting item loadouts to include new equipment.

- Add incomplete Battle class, currently only determines turn order based on entities speed stat.

+++ Dice.cs
@@ -18,9 +18,9 @@ namespace DURPSBot
             return results;
         }
 
-        public long RollTotal(int numDice, int numSides)
+        public int RollTotal(int numDice, int numSides)
         {
-            long total = 0;
+            int total = 0;
             for (int i = 0; i < numDice; i++)
             {
                 total += Rng.Next(numSides) + 1;
